requirements.txt
numpy
pillow
bitarray
psutil
smbus
spidev